requirements.txt
binance
tensorflow
sklearn
matplotlib
pandas
python-dotenv
